+++ model-gallery/README.md
@@ -1,0 +1,3 @@
+# 简介
+
+Model Gallery 是 PAI 快速开始中的一个子模块，提供了一些常用的模型，用户可以进行快速的训练和部署。本文件夹存放使用 Model Gallery 过程中可能会用到的一些样例代码。
